+++ src/resources/simulation_pause.rs
@@ -1,0 +1,3 @@
+/// Resource to insert to pause the cellular automaton simulation
+#[derive(Debug)]
+pub struct SimulationPause;